# apps/api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/sign-up": {
      "post": {
        "description": "Creates a new promoter user and returns access/refresh tokens.",
        "operationId": "Auth_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "fullName": "Jamal Salim",
                    "phone": "+254712345678",
                    "gender": "male",
                    "regionId": "nairobi-west"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "user": {
                      "id": "cmad4p0bo0000iib0i0l9e8wk",
                      "fullName": "Jamal Salim",
                      "email": null,
                      "phone": "+254712345678",
                      "uniqueCode": "P-12ab34cd",
                      "role": "promoter",
                      "gender": "male",
                      "regionId": "nairobi-west",
                      "authProvider": "credentials",
                      "requiresProfileCompletion": false,
                      "missingProfileFields": [],
                      "createdAt": "2026-05-08T18:00:00.000Z",
                      "updatedAt": "2026-05-08T18:00:00.000Z"
                    },
                    "accessToken": "<jwt>",
                    "refreshToken": "<sessionId>.<secret>",
                    "expiresIn": 900
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or request body missing"
          },
          "403": {
            "description": "Too many auth attempts from this IP"
          },
          "409": {
            "description": "User already exists"
          }
        },
        "summary": "Register a promoter account",
        "tags": ["Auth"]
      }
    },
    "/auth/signup": {
      "post": {
        "description": "Deprecated alias of /auth/sign-up; emits deprecation headers.",
        "operationId": "Auth_legacySignUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Same response as /auth/sign-up"
          }
        },
        "summary": "Deprecated signup alias",
        "tags": ["Auth"]
      }
    },
    "/auth/sign-in": {
      "post": {
        "description": "Authenticates a user and returns a fresh token pair.",
        "operationId": "Auth_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "phone": "+254712345678",
                    "uniqueCode": "P-12ab34cd",
                    "role": "promoter"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sign-in successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "user": {
                      "id": "cmad4p0bo0000iib0i0l9e8wk",
                      "fullName": "Jamal Salim",
                      "email": null,
                      "phone": "+254712345678",
                      "uniqueCode": "P-12ab34cd",
                      "role": "promoter",
                      "gender": "male",
                      "regionId": "nairobi-west",
                      "authProvider": "credentials",
                      "requiresProfileCompletion": false,
                      "missingProfileFields": []
                    },
                    "accessToken": "<jwt>",
                    "refreshToken": "<sessionId>.<secret>",
                    "expiresIn": 900
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or request body missing"
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Rate limit exceeded, phone temporarily locked, or sign-in blocked outside the authorized work area (geofence)"
          }
        },
        "summary": "Sign in with phone, code, and role",
        "tags": ["Auth"]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Deprecated alias of /auth/sign-in; emits deprecation headers.",
        "operationId": "Auth_legacySignIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Same response as /auth/sign-in"
          }
        },
        "summary": "Deprecated login alias",
        "tags": ["Auth"]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Rotates refresh token session and returns a new token pair.",
        "operationId": "Auth_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refresh successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "accessToken": "<jwt>",
                    "refreshToken": "<newSessionId>.<newSecret>",
                    "expiresIn": 900
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or request body missing"
          },
          "401": {
            "description": "Invalid, revoked, expired, or reused refresh token"
          },
          "403": {
            "description": "Refresh rate limit exceeded"
          }
        },
        "summary": "Rotate refresh token and issue new access token",
        "tags": ["Auth"]
      }
    },
    "/auth/oauth/callback": {
      "post": {
        "description": "Validates state/nonce, exchanges Google code, and returns token pair.",
        "operationId": "Auth_oauthCallback",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OauthCallbackDto"
              },
              "examples": {
                "default": {
                  "value": {
                    "provider": "google",
                    "code": "4/0AQSTgQF...",
                    "redirectUri": "http://localhost:3000/auth/google/callback",
                    "state": "eyJub25jZSI6Ii4uLiJ9.R5m..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OAuth callback successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "user": {
                      "id": "cmad4p0bo0000iib0i0l9e8wk",
                      "fullName": "Jamal Salim",
                      "email": "jamal@example.com",
                      "phone": "pending-a8e4c2f7f9d1",
                      "uniqueCode": "P-12ab34cd",
                      "role": "promoter",
                      "gender": null,
                      "regionId": null,
                      "authProvider": "google",
                      "requiresProfileCompletion": true,
                      "missingProfileFields": ["phone", "regionId"]
                    },
                    "accessToken": "<jwt>",
                    "refreshToken": "<sessionId>.<secret>",
                    "expiresIn": 900
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or malformed body"
          },
          "401": {
            "description": "Invalid OAuth state, nonce, or Google code"
          },
          "403": {
            "description": "OAuth callback rate limit exceeded"
          }
        },
        "summary": "Handle Google OAuth callback",
        "tags": ["Auth"]
      }
    },
    "/auth/oauth/google/start": {
      "post": {
        "description": "Creates state/nonce and returns Google authorization URL.",
        "operationId": "Auth_startGoogleOauth",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OauthStartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OAuth authorization payload",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "provider": "google",
                    "authorizationUrl": "https://accounts.google.com/o/oauth2/v2/auth?client_id=...&redirect_uri=...&state=...",
                    "state": "eyJub25jZSI6Ii4uLiJ9.R5m..."
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or request body missing"
          },
          "403": {
            "description": "OAuth start rate limit exceeded"
          }
        },
        "summary": "Create Google OAuth authorization URL",
        "tags": ["Auth"]
      }
    },
    "/auth/profile-completion": {
      "post": {
        "description": "Updates optional profile fields for the authenticated user.",
        "operationId": "Auth_profileCompletion",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCompletionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile completion successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "user": {
                      "id": "cmad4p0bo0000iib0i0l9e8wk",
                      "fullName": "Jamal Salim",
                      "email": "jamal@example.com",
                      "phone": "+254712345678",
                      "uniqueCode": "P-12ab34cd",
                      "role": "promoter",
                      "gender": "male",
                      "regionId": "nairobi-west",
                      "authProvider": "google",
                      "requiresProfileCompletion": false,
                      "missingProfileFields": []
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or empty payload"
          },
          "401": {
            "description": "Missing or invalid JWT token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Complete current user profile fields",
        "tags": ["Auth"]
      }
    },
    "/auth/sessions": {
      "get": {
        "description": "Returns current and historical sessions for the authenticated user.",
        "operationId": "Auth_listSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sessions fetched successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "sessions": [
                      {
                        "id": "cmadfk8m50000ii9l5g8qxh4d",
                        "userId": "cmad4p0bo0000iib0i0l9e8wk",
                        "userAgent": "Mozilla/5.0",
                        "ipAddress": "127.0.0.1",
                        "createdAt": "2026-05-08T18:00:00.000Z",
                        "updatedAt": "2026-05-08T18:00:00.000Z",
                        "expiresAt": "2026-06-07T18:00:00.000Z",
                        "revokedAt": null,
                        "replacedById": null,
                        "isCurrent": true,
                        "isActive": true
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List sessions for current user",
        "tags": ["Auth"]
      }
    },
    "/auth/sessions/revoke-all": {
      "post": {
        "description": "Revokes all sessions for the current user except current session.",
        "operationId": "Auth_revokeAllSessionsForCurrentUser",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Sessions revoked successfully"
          },
          "401": {
            "description": "Missing or invalid JWT token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke all other sessions for current user",
        "tags": ["Auth"]
      }
    },
    "/auth/sessions/revoke-user": {
      "post": {
        "description": "Admin-only endpoint to revoke all sessions for target user id.",
        "operationId": "Auth_revokeUserSessions",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeUserSessionsDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Target user sessions revoked successfully"
          },
          "401": {
            "description": "Missing or invalid JWT token"
          },
          "403": {
            "description": "Only admin users can call this endpoint"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Admin: revoke all sessions for a target user",
        "tags": ["Auth"]
      }
    },
    "/auth/sign-out": {
      "post": {
        "description": "Revokes the session represented by provided refresh token.",
        "operationId": "Auth_signOut",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignOutDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Session revoked or already inactive"
          },
          "400": {
            "description": "Validation failed or request body missing"
          }
        },
        "summary": "Revoke current refresh token session",
        "tags": ["Auth"]
      }
    },
...
  },
  "info": {
    "title": "Engaged Sales API",
    "description": "API documentation for Engaged Sales backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignUpDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Jamal Salim",
            "description": "Full name of the user"
          },
          "phone": {
            "type": "string",
            "example": "+254712345678",
            "description": "E.164 phone number"
          },
          "gender": {
            "type": "string",
            "example": "male",
            "enum": ["male", "female", "other"]
          },
          "regionId": {
            "type": "string",
            "example": "nairobi-west",
            "description": "Optional sales region id"
          }
        },
        "required": ["fullName", "phone", "gender"]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "example": "+254712345678",
            "description": "Registered phone number"
          },
          "uniqueCode": {
            "type": "string",
            "example": "P-12ab34cd",
            "description": "Unique promoter code"
          },
          "role": {
            "type": "string",
            "example": "promoter",
            "enum": ["promoter", "merchandizer", "supervisor", "admin"]
          },
          "latitude": {
            "type": "number",
            "description": "Device latitude (decimal degrees). For promoters and merchandizers: required when at least one geofence is active (send with longitude). Supervisors and admins may omit coordinates.",
            "example": -1.286389
          },
          "longitude": {
            "type": "number",
            "description": "Device longitude (decimal degrees). For promoters and merchandizers: required with latitude when geofencing is enforced. Supervisors and admins may omit.",
            "example": 36.817223
          }
        },
        "required": ["phone", "uniqueCode", "role"]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "cmadfk8m50000ii9l5g8qxh4d.6Mcpfd1xK8h6c7Iu4eD9TzLQkKkL4aY5",
            "description": "Refresh token in <sessionId>.<secret> format"
          }
        },
        "required": ["refreshToken"]
      },
      "OauthCallbackDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "example": "google",
            "enum": ["google"]
          },
          "code": {
            "type": "string",
            "example": "4/0AQSTgQF...",
            "description": "Authorization code from Google callback"
          },
          "redirectUri": {
            "type": "string",
            "example": "http://localhost:3000/auth/google/callback"
          },
          "state": {
            "type": "string",
            "example": "eyJub25jZSI6Ii4uLiJ9.R5m...",
            "description": "Signed OAuth state from start endpoint"
          },
          "latitude": {
            "type": "number",
            "description": "Device latitude. For promoters/merchandizers when geofencing is active, send with longitude. Supervisors/admins may omit.",
            "example": -1.286389
          },
          "longitude": {
            "type": "number",
            "description": "Device longitude (pair with latitude when required for field roles).",
            "example": 36.817223
          }
        },
        "required": ["provider", "code", "redirectUri", "state"]
      },
      "OauthStartDto": {
        "type": "object",
        "properties": {
          "redirectUri": {
            "type": "string",
            "example": "http://localhost:3000/auth/google/callback",
            "description": "Frontend redirect URI registered with Google OAuth"
          }
        },
        "required": ["redirectUri"]
      },
...
    }
  }
}